+++ AlchemistCode/Cards/Uncommon/PatientStrike.cs
@@ -21,6 +21,7 @@ public class PatientStrike : AlchemistCard
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_heavy_blunt"), tmpSfx: "heavy_attack.mp3").Execute(choiceContext);
+        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_heavy_blunt"),
+            tmpSfx: "heavy_attack.mp3").Execute(choiceContext);
     }
 }